refactor Server and folder - SRP. Single responsibility principle
